Reorganiza backend para funcionalidade de roles
