+++ docs/relatorios/reorganizacao-role.txt
@@ -1,0 +1,16 @@
+Relatório Técnico - Funcionalidade Role
+
+Classes movidas:
+- app/role/PapelControlador.java (anteriormente em controllers/PapelController.java)
+- app/role/PapelServico.java (anteriormente em services/PapelService.java)
+- app/role/dto/PapelDTO.java (anteriormente em app/administracao/dto/PapelDTO.java)
+- teste app/role/PapelServicoTest.java (anteriormente em services/PapelServiceTest.java)
+
+Entidades renomeadas:
+- Serviço: PapelService -> PapelServico
+- Controlador: PapelController -> PapelControlador
+
+Mocks ou dados simulados removidos: não aplicável (nenhum mock residual identificado além dos usados em testes unitários).
+
+Imports ajustados:
+- Atualização dos imports para refletir o novo pacote br.com.cloudport.servicoautenticacao.app.role e o subpacote dto.
